+++ docs/changes.md
@@ -1,6 +1,7 @@
 # 最新动态
 
 2021/12/29
+  * 修复fscript中unset的问题。
   * 允许通过 widget\_get\_prop 获取 loading 属性（感谢兆坤提供补丁）
   * 解决 combo\_box 设置 text 后，select\_index 没有跟着变化的问题（感谢兆坤提供补丁）